% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{package}
\name{dialrs-package}
\alias{dialrs}
\alias{dialrs-package}
\title{dialrs: Parse phone numbers}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Parse phone numbers using the Rust crate 'phonenumber'. Includes validation of numbers and formatting.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/schochastics/dialrs}
  \item Report bugs at \url{https://github.com/schochastics/dialrs/issues}
}

}
\author{
\strong{Maintainer}: David Schoch \email{david@schochastics.net} (\href{https://orcid.org/0000-0003-2952-4812}{ORCID})

}
