% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fake_phone}
\alias{fake_phone}
\title{A set of fake phone numbers for testing}
\format{
An object of class \code{character} of length 100.
}
\usage{
fake_phone
}
\description{
A set of fake phone numbers for testing
}
\keyword{datasets}
