% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phone.R
\name{phone_country_code}
\alias{phone_country_code}
\title{Extact country code from phone numbers}
\usage{
phone_country_code(phone)
}
\arguments{
\item{phone}{character vector of phone numbers}
}
\value{
integer vector of country codes
phone_country_code("+4916012345678")
}
\description{
Extact country code from phone numbers
}
