% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{phone_debug_print}
\alias{phone_debug_print}
\title{Print Rust internal phonenumber for debugging purposes}
\usage{
phone_debug_print(input, country)
}
\arguments{
\item{input}{character. a phone number to parse}

\item{country}{character. CLDR code}
}
\value{
nothing just used for side effects
}
\description{
Print Rust internal phonenumber for debugging purposes
}
