% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phone.R
\name{phone_is_valid}
\alias{phone_is_valid}
\title{Check if strings are valid phone numbers}
\usage{
phone_is_valid(phone, country = "")
}
\arguments{
\item{phone}{character vector of phone numbers}

\item{country}{CLDR country code. If empty, tries to parse it from \code{phone}}
}
\value{
TRUE if valid phone number, FALSE otherwise
}
\description{
Check if strings are valid phone numbers
}
\examples{
phone_is_valid(fake_phone)
}
