% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phone.R
\name{phone_iso_code}
\alias{phone_iso_code}
\title{Extract country from phone numbers in different formats}
\usage{
phone_iso_code(phone, type = "iso2")
}
\arguments{
\item{phone}{character vector of phone numbers}

\item{type}{character. one of "iso2", "iso3", or "country"}
}
\value{
character vector of country codes
phone_iso_code("+4916012345678")
}
\description{
Extract country from phone numbers in different formats
}
