% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phone.R
\name{phone_parse}
\alias{phone_parse}
\title{Parse phone numbers into specific format}
\format{
character. one of "International", "National", "RFC3966", "E.164"
}
\usage{
phone_parse(phone, country = "", format)
}
\arguments{
\item{phone}{character vector of phone numbers}

\item{country}{CLDR country code. If empty, tries to parse it from \code{phone}}
}
\value{
character vector of parsed and formatted phone numbers
phone_parse(fake_phone,format = "RFC3966")
phone_parse("016012345678",country = "DE",format = "International")
}
\description{
Parse phone numbers into specific format
}
