% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phone.R
\name{phone_type}
\alias{phone_type}
\title{Return type of phone number}
\usage{
phone_type(phone, country = "")
}
\arguments{
\item{phone}{character vector of phone numbers}

\item{country}{CLDR country code. If empty, tries to parse it from \code{phone}}
}
\value{
character vector of phone number types
}
\description{
Return type of phone number
}
